Tmp commit

+++ Server/FCOI/src/main/java/com/polus/fibicomp/fcoiDisclosure/dao/FcoiDisclosureDao.java
@@ -271,7 +271,7 @@ public interface FcoiDisclosureDao {
      * @param disclosureId
      * @return
      */
-    Integer getNumberOfSFIBasedOnDisclosureId(Integer disclosureId);
+    Long getNumberOfSFIBasedOnDisclosureId(Integer disclosureId);
 
     /**
      * This method is used to validate
